deployment/requirements.txt
streamlit==1.45.0
polars==1.29.0
pandas==2.2.3
numpy==1.24.4
scipy==1.15.2
scikit-learn==1.6.1
scikit-surprise==1.1.4
joblib==1.5.0
requests==2.32.1
python-dotenv==1.1.0
flask==3.1.0
requests